Index stickers by file_unique_id
Previously the table was indexed by file_id, but as per Telegram API
docs the value varies between bots and conversations.
As a workaround to the issue, we store both file_id and file_unique_id
in the database.

The file_unique_id is used to identify stickers when a tagger tags
(untags) a sticker, while the file_id is used by the bot to send
candidate stickers to the users for selection.